singled out api models
