+++ BussinesObject/Api/RestEntities/TestSuite/TestSuiteModel.cs
@@ -2,17 +2,21 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BussinesObject.Api.RestEntities
+namespace BussinesObject.Api.RestEntities.TestSuite
 {
-    public class CreateTestSuiteModel
+    public class TestSuiteModel
     {
-        [JsonProperty("project_id", Required = Required.Always)]
+        [JsonProperty("project_id")]
         public int ProjectId { get; set; }
 
-        [JsonProperty("name", Required = Required.Always)]
+        [JsonProperty("id")]
+        public int Id { get; set; }
+
+        [JsonProperty("name")]
         public string Name { get; set; }
 
         [JsonProperty("description")]
