+++ BussinesObject/Api/RestEntities/User/CreateUserModel.cs
@@ -1,22 +1,28 @@
 ﻿using Newtonsoft.Json;
 
-namespace BussinesObject.Api.RestEntities
+namespace BussinesObject.Api.RestEntities.User.User
 {
-    public class User
+    public class CreateUserModel
     {
         [JsonProperty("id", DefaultValueHandling = DefaultValueHandling.Ignore)]
         public int Id { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
         [JsonProperty("email")]
         public string Email { get; set; }
-        
+
         [JsonProperty("first_name")]
         public string FirstName { get; set; }
 
         [JsonProperty("last_name")]
         public string LastName { get; set; }
+
+        [JsonProperty("password")]
+        public string Password { get; set; }
+
+        [JsonProperty("password_confirmation")]
+        public string PasswordConfirmation { get; set; }
+
+        [JsonProperty("admin")]
+        public bool Admin { get; set; }
     }
 }
