--- BussinesObject/Api/RestEntities/UserResponse.cs
@@ -1,8 +1,0 @@
-﻿
-namespace BussinesObject.Api.RestEntities
-{
-    public class UserResponse
-    {
-        public User data { get; set; }
-    }
-}
