+++ BussinesObject/Api/Steps/ApiProjectSteps.cs
@@ -1,4 +1,5 @@
 ﻿using BussinesObject.Api.RestEntities;
+using BussinesObject.Api.RestEntities.Project;
 using BussinesObject.Api.Services;
 using Newtonsoft.Json;
 using NUnit.Framework;
@@ -8,13 +9,13 @@ namespace Api.BusinessObject.Steps
 {
     public class ApiProjectSteps : ProjectService
     {
-        public new Project GetProjectByCode(string code)
+        public new ProjectModel GetProjectByCode(string code)
         {
             var response = base.GetProjectByCode(code);
             Assert.IsTrue(response.StatusCode.Equals(HttpStatusCode.OK));
             Assert.IsNotNull(response.Content);
 
-            return JsonConvert.DeserializeObject<CommonResultResponse<Project>>(response.Content).Data;
+            return JsonConvert.DeserializeObject<CommonResultResponse<ProjectModel>>(response.Content).Data;
         }
     }
 }
