+++ BussinesObject/Api/Steps/ApiTestCaseSteps.cs
@@ -1,4 +1,5 @@
 ﻿using BussinesObject.Api.RestEntities;
+using BussinesObject.Api.RestEntities.TestCase;
 using BussinesObject.Api.Services;
 using BussinesObject.Api.Steps;
 using Newtonsoft.Json;
